Refine forum conversation viewport architecture
