+++ apps/www/components/school/classes/forum/conversation/item/actions.tsx
@@ -1,5 +1,6 @@
 import { ArrowTurnBackwardIcon, WinkIcon } from "@hugeicons/core-free-icons";
 import { useDisclosure } from "@mantine/hooks";
+import { captureException } from "@repo/analytics/posthog";
 import { api } from "@repo/backend/convex/_generated/api";
 import { Button } from "@repo/design-system/components/ui/button";
 import { ButtonGroup } from "@repo/design-system/components/ui/button-group";
@@ -22,6 +23,7 @@ import {
 } from "@repo/design-system/components/ui/tooltip";
 import { cn } from "@repo/design-system/lib/utils";
 import { useMutation } from "convex/react";
+import { Effect } from "effect";
 import { useTranslations } from "next-intl";
 import { useTransition } from "react";
 import { useForumSession } from "@/components/school/classes/forum/context/use-session";
@@ -80,9 +82,23 @@ export function PostItemActions({ post }: { post: ForumPost }) {
           <EmojiPicker
             className="h-80"
             onEmojiSelect={({ emoji }) => {
-              startTransition(async () => {
-                await toggleReaction({ postId: post._id, emoji });
-              });
+              startTransition(() =>
+                Effect.runPromise(
+                  Effect.tryPromise({
+                    try: () => toggleReaction({ postId: post._id, emoji }),
+                    catch: (cause) => cause,
+                  }).pipe(
+                    Effect.asVoid,
+                    Effect.catchAll((cause) =>
+                      Effect.sync(() => {
+                        captureException(cause, {
+                          source: "post-reaction-picker",
+                        });
+                      })
+                    )
+                  )
+                )
+              );
               reactionPicker.close();
             }}
           >
